This PR closes the loop to our initial “sorted runs” implementation (#2359) by explicitly tracking sorted runs and using a k-way merge to sort them.
Previously, we performed a suffix sort on each beat over the mutable range and only remembered the new sort suffix, not the concrete runs produced. 

This PR materializes sorted runs, uses an efficient k-way merge (#3226), and avoids redundant work which cuts tail latency for batches and queries.


#### Benchmarks

This is our standard benchmark with `--cache-grid=32GiB`  (and compiled with `-Dconfig_verify=true`)

##### i4i

| Metric              | Before           | This PR            |
|---------------------|------------------|------------------|
| Load accepted       | 265,856 tx/s     | **295,335 tx/s**     |
| Batch latency p99   | 174 ms           | **100 ms**           |
| Batch latency p100  | 178 ms           | **104 ms**           |
| Query latency p99   | 24 ms            | **25 ms**            |
| Query latency p100  | 48 ms            | **34 ms**            |

##### i8g

| Metric              | Before           | This PR           |
|---------------------|------------------|------------------|
| Load accepted       | 379,136 tx/s     | 402,236 tx/s     |
| Batch latency p99   | 133 ms           | 99 ms            |
| Batch latency p100  | 134 ms           | 102 ms           |
| Query latency p99   | 17 ms            | 17 ms            |
| Query latency p100  | 36 ms            | 27 ms            |

<details>

This PR explicitly tracks sorted runs and treats them as the single source of truth for whether a full table is sorted.
If exactly one run spans the full index range, the table is considered sorted.
At the end of a bar we merge all the accumulated runs with a k-way merge instead of a final sort.

*Append-only optimization*:

For append-only workloads, we extend the active run when possible.
Example: if the current run is [0..5] and we append index 6 with a key strictly greater than the key at 5, we expand the run to [0..6].

Run initialization trade-off: Starting with an initial run [0..0] looks attractive for fully sorted inserts, but it breaks down when early appends are unsorted: you end up with a tiny run [0..1]. We then proceed with a new run, because it might be sorted now, but if not, we have another tiny run. In the worst case, this produces N runs. 

To avoid that, the current approach starts with no run. We materialize the first run on the first sort_suffix call and try to expand it thereafter. Practically, if the first discovered run is [0..5], we sort that segment (even if it happens to be sorted) and then allow it to grow by one when subsequent appends remain in order. If they are not, we do not create a new sorted run; instead, we wait for the next sort_suffix call to take care of this run. This produces exactly N  runs (one beat) or fewer if they are sorted.

Alternatives considered: I tried ideas like a “sort budget” and discarding short runs. 
They increased code complexity without a measurable performance benefit.

*Possible Optimiziations*
There are minor optimizations still possible, like coalescing sorted runs: 
If two sorted runs A and B do not overlap, that is,  A.max < B.min, we could coalesce both runs into one.