@toziegler 2025-09-16
> `zig build test:unit -- tidy` is complaining about some long lines.

Fixed. Is there a way to just